% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_metriques.R
\name{calcul_aic}
\alias{calcul_aic}
\title{Fonction pour calculer l'AIC (Akaike Information Criterion)}
\usage{
calcul_aic(X, y, ll, theta)
}
\arguments{
\item{X}{: matrice des caractéristiques}

\item{y}{: vecteur des étiquettes}

\item{ll}{: log-likelyhood du modèle}

\item{theta}{: paramètres du modèle}
}
\value{
L'AIC du modèle, un nombre réel qui pénalise les modèles avec un grand nombre de paramètres.
}
\description{
Cette fonction calcule l'Akaike Information Criterion (AIC) du modèle de régression logistique.
L'AIC permet de comparer plusieurs modèles en tenant compte de la qualité de l'ajustement et du nombre de paramètres utilisés.

calculer l'AIC du modèle
l'AIC sert à comparer des modèles, en pénalisant le nb de paramètres
}
\examples{
print("hi")
}
