% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_metriques.R
\name{calcul_log_likelihood}
\alias{calcul_log_likelihood}
\title{Fonction pour calculer la log-vraisemblance}
\usage{
calcul_log_likelihood(X, y, theta)
}
\arguments{
\item{X}{: matrice des caractéristiques}

\item{y}{: vecteur des étiquettes}

\item{theta}{Matrice des paramètres (p x K), où p est le nombre de caractéristiques et K est le nombre de classes.}
}
\value{
La log-vraisemblance du modèle, un nombre réel qui mesure la qualité de l'ajustement du modèle aux données.
}
\description{
Cette fonction calcule la log-vraisemblance du modèle de régression logistique en fonction
des probabilités prédites pour chaque observation.

calculer la log-vraisemblance du modèle
}
\examples{
print("hi")
}
