% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{hessienne}
\alias{hessienne}
\title{hessian matrix}
\usage{
hessienne(X, theta)
}
\arguments{
\item{X}{: data matrix (n x p)}

\item{theta}{: matrix of regression coefficients (p x K)}
}
\value{
hessian matrix (p x p)
}
\description{
Computes the hessian matrix
}
