% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_x.R
\name{prepare_x}
\alias{prepare_x}
\title{Préparer la matrice X à la prise en compte des variables mixtes}
\usage{
prepare_x(X)
}
\arguments{
\item{X}{: data.frame contenant les données}
}
\value{
Une matrice \code{X} préparée, avec les variables qualitatives encodées, les variables quantitatives normalisées,
et une colonne d'intercept ajoutée (si ce n'est pas déjà présent).

X encodé et normalisé
}
\description{
Cette fonction prépare la matrice X en encodant les variables qualitatives (s'il y a plus de 2 modalités),
en normalisant les variables quantitatives et en ajoutant une colonne d'intercept (si nécessaire).

encode variables quali et normalise les quanti + intercept
}
\examples{
print("Hi")
}
