% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{print_coeffs}
\alias{print_coeffs}
\title{Afficher un résumé des coefficients}
\usage{
print_coeffs(dict_coeff)
}
\arguments{
\item{dict_coeff}{: liste de dataframes contenant
coeffs, erreurs std, z-scores et p-values}
}
\value{
Affiche les coefficients sous forme de texte.

coefficients de la régression
}
\description{
Cette fonction affiche un résumé des coefficients de régression, y compris les erreurs standards,
les z-scores, les p-values et les niveaux de significativité pour chaque classe de la régression.

afficher les coefficients de la régression
}
\examples{
print("Hi")
}
