% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_multinomiale.R
\name{reg_multinomiale}
\alias{reg_multinomiale}
\title{Multinomial logistic regression}
\usage{
reg_multinomiale(X, y, nb_iters, alpha, penalty, lambda, l1_ratio)
}
\arguments{
\item{X}{(data.frame) Predictor variables (features) for the model.}

\item{y}{(vector) Target variable (labels) for training the model.}

\item{nb_iters}{(integer) Number of iterations for the gradient descent.}

\item{alpha}{(numeric) Learning rate for the gradient descent.}

\item{penalty}{(character) Type of regularization: l1=lasso, l2=ridge, elasticnet.}

\item{lambda}{(numeric) Regularization parameter.}

\item{l1_ratio}{(numeric) The ElasticNet mixing parameter, with 0 <= l1_ratio <= 1. Only used if penalty is 'elasticnet'.}
}
\value{
(matrix) Optimized parameters for each class.
}
\description{
Performs multinomial logistic regression to find optimal parameters.
}
