% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descente_gradient.R
\name{sigmoid}
\alias{sigmoid}
\title{Fonction sigmoïde}
\usage{
sigmoid(z)
}
\arguments{
\item{z}{: vecteur ou matrice}
}
\value{
Un vecteur ou une matrice de même taille que \code{z}, où chaque élément correspond à la valeur de la fonction sigmoïde appliquée à l'élément correspondant de \code{z}.

sigmoïde de z (valeur comprise entre 0 et 1)
}
\description{
Calcule la fonction sigmoïde de l'entrée \code{z}. La sigmoïde est une fonction qui transforme
les valeurs d'entrée en un résultat compris entre 0 et 1.
}
\examples{
print("hi")
}
