% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_x.R
\name{type_variable}
\alias{type_variable}
\title{Déterminer les variables qualitatives et quantitatives}
\usage{
type_variable(df)
}
\arguments{
\item{df}{: data.frame contenant les données}
}
\value{
Une liste contenant deux éléments :
- \code{qualitatives} : un vecteur des noms des variables qualitatives
- \code{quantitatives} : un vecteur des noms des variables quantitatives

liste contenant les variables qualitatives et quantitatives
}
\description{
Cette fonction permet de différencier les variables qualitatives (facteurs et caractères)
et quantitatives (numériques et entières) dans un data.frame.
}
\examples{
print("Hi")
}
