% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_metriques.R
\name{calcul_aic}
\alias{calcul_aic}
\title{Function to Calculate AIC (Akaike Information Criterion)}
\usage{
calcul_aic(X, y, ll, theta)
}
\arguments{
\item{X}{Feature matrix}

\item{y}{Label vector (n x 1)}

\item{ll}{Log-likelihood of the model}

\item{theta}{Matrix of regression coefficients (p x K),
p = number of features and K = number of classes.}
}
\value{
The AIC of the model
}
\description{
This function calculates the Akaike Information Criterion (AIC).
The AIC allows comparing several models with each other.
It takes into account the quality of the fit and the number of parameters used.
}
