% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_metriques.R
\name{calcul_log_likelihood}
\alias{calcul_log_likelihood}
\title{Function to Calculate Log-Likelihood}
\usage{
calcul_log_likelihood(X, y, theta)
}
\arguments{
\item{X}{Feature matrix (n x p),
n = number of individuals and p = number of variables}

\item{y}{Label vector (n x 1)}

\item{theta}{Matrix of regression coefficients (p x K),
p = number of features and K = number of classes.}
}
\value{
The log-likelihood of the model
}
\description{
This function calculates the log-likelihood of the model
based on the predicted probabilities for each observation.
}
