% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{calcul_p_values}
\alias{calcul_p_values}
\title{Calculate p-values of regression coefficients}
\usage{
calcul_p_values(X, theta)
}
\arguments{
\item{X}{matrix of input data}

\item{theta}{matrix of regression coefficients}
}
\value{
list of dataframes containing:
coefficients, standard errors, z-scores, and p-values
for each class of the regression
}
\description{
Calculate p-values of regression coefficients
}
