% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_proba.R
\name{predict_proba}
\alias{predict_proba}
\title{Predicts class membership probabilities}
\usage{
predict_proba(X, theta)
}
\arguments{
\item{X}{matrix or data.frame of individuals' features
row = an individual, column = a feature}

\item{theta}{matrix of learned coefficients
column = a class, row = a feature}
}
\value{
matrix of class membership probabilities for each individual
each row corresponds to an individual and each column to a class
}
\description{
Predicts the probabilities of individuals belonging to classes.
Uses the scores obtained by multiplying X with theta,
then applies the softmax function to obtain the probabilities
}
