% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_x.R
\name{prepare_x}
\alias{prepare_x}
\title{Prepare the matrix X to account for mixed variables}
\usage{
prepare_x(X)
}
\arguments{
\item{X}{: data.frame containing the data}
}
\value{
encoded and normalized X
}
\description{
Prepares the matrix X by encoding qualitative variables (if more than 2 modalities),
normalizing quantitative variables, and adding an intercept column.
}
