% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{print_coeffs}
\alias{print_coeffs}
\title{Display a summary of coefficients}
\usage{
print_coeffs(dict_coeff)
}
\arguments{
\item{dict_coeff}{: list of dataframes containing
coefficients, standard errors, z-scores, and p-values}
}
\value{
regression coefficients
}
\description{
Display the regression coefficients
}
