% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_x.R
\name{type_variable}
\alias{type_variable}
\title{Determines qualitative and quantitative variables}
\usage{
type_variable(df)
}
\arguments{
\item{df}{: data.frame containing the data}
}
\value{
list containing qualitative and quantitative variables
}
\description{
differentiates between qualitative variables (factors and characters)
and quantitative variables (numeric and integer) in a data.frame
WARNING : if the variables are not correctly encoded this won't work
}
